+++ src/app/View/Modal/Modal.tsx
@@ -10,7 +10,7 @@ const ContentModal = dynamic(() => import("../Main/ContentModal"), {
 const FooterItemModal = dynamic(() => import("../Footer/FooterItemModal"), {
   ssr: false,
 });
-import ArticleEditor from "../Drawer/ArticleEditor/ArticleEditor";
+import ArticleEditor from "../Drawer/ArticleEditor/view/ArticleEditor";
 import FooterItemEditor from "../Drawer/ItemEditor/FooterItemEditor";
 import { FeedbackForm } from "../Drawer/FeedbackForm";
 import { ManageTheme } from "../Drawer/ManageTheme/ManageTheme";
